+++ pytsite/geo/_widget.py
@@ -29,10 +29,8 @@ class Location(_widget.Abstract):
 
         super().__init__(uid, **kwargs)
 
-        # Autodetect location
         self._autodetect = kwargs.get('autodetect', False)
-
-        self._js_module('pytsite-geo-widget-location')
+        self._js_module = 'pytsite-geo-widget-location'
         self._css += ' widget-geo-location'
 
         # Validation rule for 'required' widget