+++ pytsite/widget/res/assets/js/score.js
@@ -1,4 +1,6 @@
-define(['jquery'], function ($) {
+define(['jquery', 'assetman'], function ($, assetman) {
+    assetman.loadCSS('pytsite.widget@css/score.css');
+
     return function (widget) {
         var switches = widget.em.find('.switch');
         var input = widget.em.find('input');